+++ pom.xml
@@ -6,6 +6,7 @@
     <packaging>jar</packaging>
     <version>1.0-SNAPSHOT</version>
     <name>mqtt</name>
+    <url>http://maven.apache.org</url>
     <repositories>
         <repository>
             <id>paho-mqtt-client</id>
@@ -13,7 +14,6 @@
             <url>https://repo.eclipse.org/content/repositories/paho-releases/</url>
         </repository>
     </repositories>
-    <url>http://maven.apache.org</url>
     <dependencies>
         <dependency>
             <groupId>org.eclipse.paho</groupId>
